requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
html5lib==0.9999999
Keras==2.0.5
Markdown==2.2.0
matplotlib==2.0.2
mock==2.0.0
numpy
olefile==0.44
openmdao==1.7.3
pbr==3.1.1
Pillow==4.1.1
progressbar2
protobuf==3.3.0
psutil
py==1.4.34
pyparsing==2.2.0
pytest==3.1.2
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
sampyl-mcmc
scipy==0.19.1
si-prefix==0.4.post10
six==1.10.0
subprocess32==3.2.7
tabulate==0.7.7
torch
torchvision
Werkzeug==0.12.2
